Fixed Names, Corrected Errors, added notes
