+++ src/components/Cohort.js
@@ -1,27 +1,35 @@
 import './Cohort.css'
-export default function Cohort({student}){
-    const cohortList =[
+export default function Cohort({student, setCohorts}){
+    const cohortList = [
         ...new Set(student
             .map((students) => students.cohort.cohortCode))
     ]
-    //CREDIT TO BRANDON FOR USE
 
+    //New set is making sure I don't have to repeat this code.
+ 
     const space = (e) => {
         return e.split('').slice(0, -4).join('') + " " + e.slice(1).slice(-4)
     }
-
-
     //THIS IS ADDING THE SPACE IN THE DATA. FIND A BETTER WAY TO DO THIS 
+
+    const handleClick = (e) =>{
+        const {value} = e.target
+            setCohorts(value)
+    }
     return(
         <div className='classDate'>
             <h2 className="other">Choose a Class by Start Date</h2>
             <ul>
-                {cohortList.map((cohorts, index)=>(
-                    <li key={index}>
-                    <button onClick={cohorts}>{space(cohorts)}
-                    </button>
-                    </li>
-                ))}
+                <button value="All Students" onClick={handleClick}>All Students</button>
+                {cohortList.map((cohort)=>{
+                    return(
+                        <div>
+                            <button value={cohort} onClick={handleClick}>
+                                {space(cohort)}
+                            </button>
+                        </div>
+                    )
+                })}
             </ul>
         </div>
     )
